+++ malawi/inputHandlers/accountNumberInputHandler.js
@@ -1,23 +1,22 @@
 var getClient = require('../utils/getClient');
-var buybackTransactions =  require('../../buyback-transactions/buyBackTransactions');
-var accountNumberInputHandler = 'account_number';
+var onAccountNumberValidated = require('../utils/onAccountNumberValidated');
+var translations = require('../translations/index');
+var translator = require('../../utils/translator/translator');
+var accountNumberInputHandler = 'mw_account_number_handler';
 
 module.exports = {
     handlerName: accountNumberInputHandler,
-    handler: function(input){
-        var account_number = input.replace(/\D/g, '');
-
-        var client = getClient(account_number);
-        if(client.error_message){
-            sayText(client.error_message);
-            promptDigits('account_number', {
-                'submitOnHash': false,
-                'maxDigits': 8,
-                'timeout': 10,
-            });
-        } else {
-            // spin the buyback implementation
-            buybackTransactions.start(client.client);
-        } 
+    getHandler: function(lang) {
+        return function(input){
+            var account_number = input.replace(/\D/g, '');
+            var getMessage = translator(translations, lang);
+            var client = getClient(account_number);
+            if(client.error_message){
+                global.sayText(getMessage('invalid_account', {}, lang));
+                global.promptDigits('account_number');
+            } else {
+                onAccountNumberValidated(lang, client);
+            } 
+        };
     }
 };