+++ malawi/inputHandlers/inputHandlers.js
@@ -1,5 +1,0 @@
-var accountNumberInputHandler = require('./accountNumberInputHandler');
-
-module.exports = {
-    accountNumberInputHandler: accountNumberInputHandler.handler,
-};